requirements.txt
fastapi
uvicorn[standard]
jinja2
openai
google-genai
pinecone
sqlalchemy
tavily-python
requests
python-dotenv
pydantic
bcrypt
PyJWT